# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Crew Intelligence Platform — AI Services API
  description: |
    REST API for the CIP Python microservices layer.
    Provides Excel parsing, legality engine, AI assistant (Najm), smart ranking, and auto-bid services.
    
    **Authentication:** All endpoints require a valid Firebase ID token passed as `Authorization: Bearer <token>`.
    Service-to-service calls use a signed service account JWT.
    
    **Base URL (production):** `https://cip-python-xxxx.run.app`
    **Base URL (development):** `http://localhost:8080`
  version: 1.0.0
  contact:
    name: CIP Support
    email: support@cip.app
servers:
  - url: https://cip-python-xxxx.run.app
    description: Production (Google Cloud Run — me-central1)
  - url: http://localhost:8080
    description: Local development
security:
  - FirebaseAuth: []
...
paths:
...
  /v1/intelligence/upload:
    post:
      tags: [PDF Intelligence]
      summary: Upload a schedule PDF for analysis
      description: |
        Multipart upload. Identity is taken from the Firebase ID token — a
        user_id form field, if sent, is ignored for non-service callers
        (F11). Files above 20 MB are rejected (413); `period` must look like
        MMM-YYYY and `year` must be plausible. Processing is asynchronous —
        poll the status endpoint.
      parameters:
        - name: user_id
          in: query
          required: true
          schema: { type: string }
          description: Ignored for non-service callers — identity is pinned
            to the Firebase token uid (F11).
        - name: period
          in: query
          required: true
          schema: { type: string, example: JUN-2026 }
        - name: year
          in: query
          required: true
          schema: { type: integer, example: 2026 }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary, description: PDF ≤ 20 MB }
      responses:
        "200":
          description: Upload accepted; analysis queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  uploadId: { type: string }
                  status: { type: string, example: queued }
        "400": { description: Non-PDF payload, bad period format, or implausible year }
        "413": { description: File exceeds the 20 MB limit }
        "503": { description: Upload tracking unavailable — retry }
  /v1/intelligence/upload/{upload_id}/status:
    get:
      tags: [PDF Intelligence]
      summary: Poll an upload's processing status
      description: Owner-only — the requesting token must match the upload's user.
      parameters:
        - name: upload_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Current status
          content:
            application/json:
              schema:
                type: object
                properties:
                  uploadId: { type: string }
                  status:
                    type: string
                    enum: [queued, processing, complete, failed, not_found, error]
                  lineId: { type: string, nullable: true }
                  error: { type: string, nullable: true }
        "403": { description: Not the upload owner }

  # ─── AI Assistant ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    FirebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
...